test(pg): fold BC array regression back into array.test.ts under trails describe
A separate array.trails.test.ts would be a second pg_arrays DDL owner,
racing the mirror on the shared multi-fork PG database (single
PG_TEST_URL, no per-worker DB). Keep the test in the single-owner file
under a distinct non-mirror describe so it stays out of the
PostgresqlArrayTest test:compare mapping without a second table owner.
