+++ packages/activerecord/src/adapters/postgresql/array.test.ts
@@ -666,4 +666,33 @@ describeIfPg("PostgreSQLAdapter", () => {
       expect((record.timestamps[0] as Temporal.Instant).epochNanoseconds % 1000000n).toBe(123000n);
     });
   });
+
+  // Trails-only coverage with no counterpart in Rails' PostgresqlArrayTest.
+  // Kept in this file (reusing its single `pg_arrays` owner) rather than a
+  // `*.trails.test.ts` sibling: a second file dropping/creating `pg_arrays` in
+  // beforeAll/afterAll would race the mirror on the shared PG database (multi-
+  // fork, single PG_TEST_URL). The distinct describe name keeps it out of the
+  // `PostgresqlArrayTest` → Rails test:compare mapping.
+  describe("array datetime inline-quoting (trails)", () => {
+    it("inlines a proleptic-year datetime[] element as a quoted_date BC literal", async () => {
+      // Regression guard for the inline INSERT path (base.ts create → adapter
+      // `quote` → encode_array → type_cast → quoted_date). A proleptic year <= 0
+      // only round-trips through the " BC" literal; the pre-fix ISO fallthrough
+      // (`{-000042-03-15T...Z}`) is not valid PG array input, so this pins the
+      // routing rather than a form PG happens to also accept (ISO datetimes do).
+      class PgArrays extends Base {
+        static tableName = "pg_arrays";
+        static {
+          this.attribute("id", "integer");
+        }
+      }
+      await PgArrays.loadSchema();
+      // Temporal proleptic year -42 == 43 BC.
+      const bc = Temporal.Instant.from("-000042-03-15T12:34:56.123456Z");
+      const record = await (PgArrays as any).create({ datetimes: [bc] });
+      await record.reload();
+      expect(record.datetimes).toHaveLength(1);
+      expect((record.datetimes[0] as Temporal.Instant).epochNanoseconds).toBe(bc.epochNanoseconds);
+    });
+  });
 });
